+++ src/tables/cff/charset.rs
@@ -2,26 +2,6 @@ use crate::GlyphId;
 use crate::parser::{Stream, FromData, LazyArray16};
 use super::StringId;
 
-/// The Standard Encoding as defined in the Adobe Technical Note #5176 Appendix B.
-pub const STANDARD_ENCODING: [u8;256] = [
-      0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-      0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-      1,   2,   3,   4,   5,   6,   7,   8,   9,  10,  11,  12,  13,  14,  15,  16,
-     17,  18,  19,  20,  21,  22,  23,  24,  25,  26,  27,  28,  29,  30,  31,  32,
-     33,  34,  35,  36,  37,  38,  39,  40,  41,  42,  43,  44,  45,  46,  47,  48,
-     49,  50,  51,  52,  53,  54,  55,  56,  57,  58,  59,  60,  61,  62,  63,  64,
-     65,  66,  67,  68,  69,  70,  71,  72,  73,  74,  75,  76,  77,  78,  79,  80,
-     81,  82,  83,  84,  85,  86,  87,  88,  89,  90,  91,  92,  93,  94,  95,   0,
-      0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-      0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-      0,  96,  97,  98,  99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110,
-      0, 111, 112, 113, 114,   0, 115, 116, 117, 118, 119, 120, 121, 122,   0, 123,
-      0, 124, 125, 126, 127, 128, 129, 130, 131,   0, 132, 133,   0, 134, 135, 136,
-    137,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,   0,
-      0, 138,   0, 139,   0,   0,   0,   0, 140, 141, 142, 143,   0,   0,   0,   0,
-      0, 144,   0,   0,   0, 145,   0,   0, 146, 147, 148, 149,   0,   0,   0,   0,
-];
-
 /// The Expert Encoding conversion as defined in the Adobe Technical Note #5176 Appendix C.
 #[cfg(feature = "glyph-names")]
 const EXPERT_ENCODING: &[u16] = &[
@@ -101,12 +81,6 @@ pub(crate) enum Charset<'a> {
 }
 
 impl Charset<'_> {
-    pub fn glyph_index(&self, code_point: u8) -> Option<GlyphId> {
-        let sid = STANDARD_ENCODING[usize::from(code_point)];
-        let sid = StringId(u16::from(sid));
-        self.sid_to_gid(sid)
-    }
-
     pub fn sid_to_gid(&self, sid: StringId) -> Option<GlyphId> {
         if sid.0 == 0 {
             return Some(GlyphId(0));